service-workers/service-worker/postmessage-to-client.https.html

<!DOCTYPE html>
<title>Service Worker: postMessage to Client</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/get-host-info.sub.js"></script>
<script src="resources/test-helpers.sub.js"></script>
<script>
promise_test(async t => {
  const SCRIPT = 'resources/postmessage-to-client-worker.js';
  const SCOPE = 'resources/blank.html';
  let contentWindow;

  const registration =
      await service_worker_unregister_and_register(t, SCRIPT, SCOPE);
  await wait_for_state(t, registration.installing, 'activated');
  const frame = await with_iframe(SCOPE);

  t.add_cleanup(async() => {
    if (frame)
      frame.remove();
    if (registration)
      await registration.unregister();
  });

  let event = await new Promise(resolve =>{
    contentWindow = frame.contentWindow;
    contentWindow.navigator.serviceWorker.onmessage = resolve;
    contentWindow.navigator.serviceWorker.controller.postMessage('ping');
  });

  const message = event.data;
  assert_equals(event.constructor, contentWindow.MessageEvent,
                'message events should use MessageEvent interface.');
  assert_equals(event.type, 'message', 'type should be "message".');
  assert_equals(event.bubbles, false, 'message events should not bubble.');
  assert_equals(event.cancelable, false,
                'message events should not be cancelable.');
  assert_equals(event.origin, location.origin,
                'origin of message should be origin of Service Worker.');
  assert_equals(event.lastEventId, '',
                'lastEventId should be an empty string.');
  assert_equals(event.source.constructor, contentWindow.ServiceWorker,
                'source should use ServiceWorker interface.');
  assert_equals(event.source, contentWindow.navigator.serviceWorker.controller,
                'source should be the service worker that sent the message.');
  assert_equals(event.ports.length, 0, 'ports should be an empty array.');
  assert_equals(message, 'Sending message via clients');

  event = await new Promise(resolve => {
          contentWindow.navigator.serviceWorker.onmessage = resolve;
  });
  assert_equals(event.data, 'quit');

  }, 'postMessage from ServiceWorker to Client.');
</script>
